requirements.txt
# python==3.8.19
matplotlib==3.7.5
wandb==0.12.1
pillow==8.4.0 # 9.3.0
numpy==1.21.0
h5py==3.3.0
# lightly==1.1.21
gdown==3.13.0
pyrender==0.1.45
protobuf==3.20.0
vedo==2023.4.4
pymeshlab==2022.2.post4
easydict==1.13
# open3d==0.18.0 #先注释掉
pykeops==2.2.3
einops==0.8.0
torch-geometric==2.5.3
timm==1.0.3
addict==2.4.0
spconv-cu117==2.3.6
# torch-scatter==2.1.1+pt113cu117
# torch-sparse==0.6.17+pt113cu117
# torch-cluster==1.6.1+pt113cu117
# torch-spline-conv==1.2.2+pt113cu117
# flash-attn==2.5.9.post1
# torch-points-kernels==0.6.10
# torch-points3d==1.3.0

# https://pytorch-geometric.readthedocs.io/en/2.3.0/notes/installation.html
# windows run:
# pip install torch_scatter torch_sparse torch_cluster torch_spline_conv -f https://data.pyg.org/whl/torch-1.13.0+cu117.html
# linux run:
# pip install pyg_lib torch_scatter torch_sparse torch_cluster torch_spline_conv -f https://data.pyg.org/whl/torch-1.13.0+cu117.html

# pip install torch==1.13.1+cu117 torchvision==0.14.1+cu117 torchaudio==0.13.1 --extra-index-url
# 下载用pip install torch==1.13.1+cu117 torchvision==0.14.1+cu117 torchaudio==0.13.1 --extra-index-url https://download.pytorch.org/whl/cu117
